requirements.txt
pandas
requests
pymongo
python-dotenv
beautifulsoup4
lxml
transformers[torch]
sentence-transformers
keybert
google-generativeai